Implemented ActiveBoard
